move state structs and transitions to new systems module
