+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move set_window_icon temporary system to PreStartup schedule
 - Move configure_physics_engine to GameCreate state
 - Move system sets to new systems module
+- Move state structs and transitions to new systems module
 
 ## [0.0.19] - 2023-10-31
